raycast ok normalenent

+++ incl/wolf3d.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   wolf3d.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ibouabda <ibouabda@student.42.fr>          +#+  +:+       +#+        */
+/*   By: idris <idris@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/30 13:45:07 by ibouabda          #+#    #+#             */
-/*   Updated: 2019/12/12 17:07:06 by ibouabda         ###   ########.fr       */
+/*   Updated: 2019/12/13 10:44:48 by idris            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -124,7 +124,7 @@ typedef struct	s_env
 	int				midx;
 	int				midy;
 	int				dist;
-	float			ang;
+	double			ang;
 	int				rot;
 	char			**dbtab;
 	t_image			tex;
